clean up publish yaml refactor deploys test prod deploy
