+++ .github/workflows/publish.yaml
@@ -5,13 +5,6 @@ on:
     tags:
       - "*"
 
-# on: [push]
-#     tags:
-#       - "*"
-# on:
-#   release:
-#     types: [published]
-
 jobs:
   buildAndTest:
     name: Build and Test
@@ -61,14 +54,6 @@ jobs:
         uses: borales/actions-yarn@v4
         with:
           cmd: install
-      # - name: Build Project
-      #   uses: borales/actions-yarn@v4
-      #   with:
-      #     cmd: build
-      # - name: Run Tests
-      #   uses: borales/actions-yarn@v4
-      #   with:
-      #     cmd: test
       - name: Prepack
         uses: borales/actions-yarn@v4
         with:
@@ -87,8 +72,6 @@ jobs:
     needs: prepack
     runs-on: ubuntu-latest
     steps:
-      # - uses: actions/checkout@v4
-
       - name: Download build artifact from prepack job
         uses: actions/download-artifact@v4
         with:
@@ -104,18 +87,6 @@ jobs:
         uses: borales/actions-yarn@v4
         with:
           cmd: install
-      # - name: Install dependencies
-      #   uses: borales/actions-yarn@v4
-      #   with:
-      #     cmd: versionPackage #need to reset the version to the latest tag or it'll fail to deploy
-      # - name: Build Project
-      #   uses: borales/actions-yarn@v4
-      #   with:
-      #     cmd: build
-      # - name: Run Tests
-      #   uses: borales/actions-yarn@v4
-      #   with:
-      #     cmd: test
       - name: Deploy Contracts
         uses: borales/actions-yarn@v4
         with:
@@ -161,18 +132,6 @@ jobs:
         uses: borales/actions-yarn@v4
         with:
           cmd: install
-      # - name: Build Project
-      #   uses: borales/actions-yarn@v4
-      #   with:
-      #     cmd: build
-      # - name: Run Tests
-      #   uses: borales/actions-yarn@v4
-      #   with:
-      #     cmd: test
-      # - name: Prepack
-      #   uses: borales/actions-yarn@v4
-      #   with:
-      #     cmd: prepack
       - name: Publish to NPM
         uses: borales/actions-yarn@v4
         with:
