+++ scripts/deloyProtocol.ts
@@ -2,7 +2,10 @@ import { ethers, run } from 'hardhat'
 import { updateAddresses } from './deploy/addresses'
 import { Signer } from 'ethers'
 import { Deployer, Inbox, Prover } from '../typechain-types'
+import { networks as mainnetNetworks } from '../config/mainnet/config'
+import { networks as sepoliaNetworks } from '../config/testnet/config'
 import { Address, Hex } from 'viem'
+import { isZeroAddress } from './utils'
 export const ALCHEMY_API_KEY = process.env.ALCHEMY_API_KEY || ''
 
 export type DeployNetwork = {
@@ -23,6 +26,95 @@ export type ProtocolDeploy = {
   inboxAddress: Hex
   hyperProverAddress: Hex
   initialSalt: string
+}
+
+export async function deployProtocol(protocolDeploy: ProtocolDeploy, deployNetwork: DeployNetwork, solver: Hex, proverConfig: any, isSolvingPublic: boolean = true) {
+  const networkName = deployNetwork.name
+  const salt = ethers.keccak256(ethers.toUtf8Bytes(protocolDeploy.initialSalt))
+  const [deployer] = await ethers.getSigners()
+  console.log('Deploying contracts with the account:', deployer.address)
+  if (process.env.DEPLOY_CI === 'true') {
+    console.log('Deploying for CI')
+  }
+
+  const singletonDeployer = await ethers.getContractAt(
+    'Deployer',
+    '0xfc91Ac2e87Cc661B674DAcF0fB443a5bA5bcD0a3',
+  )as any as Deployer
+
+  console.log('gasLimit:', deployNetwork.gasLimit)
+
+  console.log(`***************************************************`)
+  console.log(`** Deploying contracts to ${networkName} network **`)
+  console.log(`***************************************************`)
+
+  if (isZeroAddress(protocolDeploy.proverAddress)) {
+    await deployProver(salt, deployNetwork, singletonDeployer, proverConfig)
+  }
+
+  if (isZeroAddress(protocolDeploy.intentSourceAddress)) {
+    protocolDeploy.intentSourceAddress = await deployIntentSource(
+      deployNetwork,
+      salt,
+      singletonDeployer,
+    )
+  }
+
+  if (isZeroAddress(protocolDeploy.inboxAddress)) {
+    protocolDeploy.inboxAddress = await deployInbox(
+      deployNetwork,
+      deployer,
+      isSolvingPublic,
+      [solver],
+      salt,
+      singletonDeployer,
+    )
+  }
+
+  if (
+    isZeroAddress(protocolDeploy.hyperProverAddress) &&
+    !isZeroAddress(protocolDeploy.inboxAddress)
+  ) {
+    protocolDeploy.hyperProverAddress = await deployHyperProver(
+      deployNetwork,
+      protocolDeploy.inboxAddress,
+      salt,
+      singletonDeployer,
+    )
+  }
+}
+
+export function getDeployNetwork(networkName: string): DeployNetwork {
+//mainnet
+  switch (networkName) {
+    case 'base':
+      return  mainnetNetworks.base
+    case 'optimism':
+      return mainnetNetworks.optimism
+    case 'helix':
+      return mainnetNetworks.helix
+    case 'arbitrum':
+      return mainnetNetworks.arbitrum
+    case 'mantle':
+      return mainnetNetworks.mantle
+  }
+
+  //sepolia
+  switch (networkName) {
+    case 'baseSepolia':
+      return sepoliaNetworks.baseSepolia
+    case 'optimismSepolia':
+      return sepoliaNetworks.optimismSepolia
+    case 'optimismSepoliaBlockscout':
+      return sepoliaNetworks.optimismSepolia
+    case 'ecoTestnet':
+      return sepoliaNetworks.ecoTestnet
+    case 'arbitrumSepolia':
+      return sepoliaNetworks.arbitrumSepolia
+    case 'mantleSepolia':
+      return sepoliaNetworks.mantleSepolia
+  }
+  throw new Error('Network not found')
 }
 
 export async function deployProver(
@@ -120,7 +212,7 @@ export async function deployInbox(
     contractName,
     inboxAddress,
     inboxOwnerSigner,
-  )
+  ) as any as Inbox
 
   await inbox
     .connect(inboxOwnerSigner)
